Add first draft of webmap quickstart tutorial.
Add a new tutorial index that collects tutorials and how-to's about
using the API.

Add a tutorial on using Mapbox GL JS and the API to make a simple
webapp displaying metadata and an interactive map for a specific Tree
Cover Loss dataset.

Modify `screen.css.scss` stylesheet to enable code blocks in the middle
column of slatedocs.

Changes to be committed:
  new file:   source/includes/tutorial/webdev_mapbox_quickstart/_quickstart_webmap.md
  new file:   source/tutorials.html.md
  modified:   source/stylesheets/screen.css.scss