Create geometry analysis tutorial (#153)
Create a new tutorial for geometry analysis, which covers working with
vector tiles, the query endpoint, and shows a practical example of how
datasets are likely to be created, transformed, and linked together
across different providers to make a more ready-to-use system.

In the future this tutorial will need to be updated with new dataset IDs as they are made available. Probably every quarter or so.